Add WeatherDescriptionDto.toModel() extension funtion

+++ app/src/main/java/com/uxstate/skycast/domain/model/WeatherDescription.kt
@@ -1,3 +1,3 @@
 package com.uxstate.skycast.domain.model
 
-data class WeatherDescription(val id:Long, val main:String, val description:String, val icon:String)
+data class WeatherDescription(val id:Int, val main:String, val description:String, val icon:String)
